requirements.txt
nibabel==2.3.1
six==1.12.0
numpy==1.16.2
chainer==6.2.0
scikit_image==0.14.2
imageio==2.4.1
cupy_cuda90==6.2.0
fire==0.2.1
pyyaml==3.13
scikit_learn==0.22.1
mpi4py==3.0.0
SimpleITK==1.2.0
pillow==4.3.0
opencv-python==4.1.0.25